buttons group fix
